aem-automation-agent/requirements.txt
playwright==1.40.0
openai==1.3.8
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2
asyncio-throttle==1.0.2
flask==2.3.3
